@@ -5,19 +5,17 @@
   "main": "webpack.config.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "webpack-dev-server --progress --colors",
-    "build": "webpack --optimize-minimize"
+    "build": "webpack"
   },
   "author": "Patrick Arlt",
   "license": "Apache-2.0",
   "dependencies": {
-    "esri-leaflet": "^2.1.1",
-    "esri-leaflet-geocoder": "^2.2.6",
-    "leaflet": "^1.2.0"
+    "esri-leaflet": "^2.1.3",
+    "esri-leaflet-geocoder": "^2.2.8"
   },
   "devDependencies": {
-    "webpack": "^3.5.5",
-    "webpack-dev-server": "^2.7.1"
+    "webpack": "^4.0.1",
+    "webpack-cli": "^2.0.9"
   },
   "repository": {
     "type": "git",
