@@ -5,6 +5,8 @@ module.exports = {
   output: {
     filename: 'bundle.js',
     path: path.resolve(__dirname, 'dist')
-  }
+  },
+  mode: 'production',
+  plugins: []
 };
 
